Change heights to use u8 instead of u64, since HEIGHT <= 24

+++ src/internal/hash.rs
@@ -112,7 +112,7 @@ impl Hash {
 
     #[inline]
     pub(crate) fn node(
-        height: u64,
+        height: u8,
         Hash(a): Hash,
         Hash(b): Hash,
         Hash(c): Hash,
